--- langchain/cli/conf/nginx.conf
@@ -1,16 +1,0 @@
-server {
-    listen       80;
-    server_name  localhost;
-    error_log /var/log/nginx/error.log warn;
-
-    location / {
-        root   /usr/share/nginx/html;
-        index  index.html index.htm;
-        try_files $uri $uri/ /index.html;
-    }
-
-    error_page   500 502 503 504  /50x.html;
-    location = /50x.html {
-        root   /usr/share/nginx/html;
-    }
-}
