--- langchain/cli/docker-compose.ngrok.yaml
@@ -1,17 +1,0 @@
-version: '3'
-services:
-  ngrok:
-    image: ngrok/ngrok:latest
-    restart: unless-stopped
-    command:
-      - "start"
-      - "--all"
-      - "--config"
-      - "/etc/ngrok.yml"
-    volumes:
-      - ./ngrok_config.yaml:/etc/ngrok.yml
-    ports:
-      - 4040:4040
-  langchain-backend:
-    depends_on:
-      - ngrok
