--- langchain/cli/docker-compose.yaml
@@ -1,49 +1,0 @@
-version: '3'
-services:
-  langchain-frontend:
-    image: langchain/${_LANGCHAINPLUS_IMAGE_PREFIX-}langchainplus-frontend:latest
-    ports:
-      - 80:80
-    environment:
-      - REACT_APP_BACKEND_URL=http://localhost:1984
-    depends_on:
-      - langchain-backend
-    volumes:
-      - ./conf/nginx.conf:/etc/nginx/default.conf:ro
-    build:
-      context: frontend-react/.
-      dockerfile: Dockerfile
-  langchain-backend:
-    image: langchain/${_LANGCHAINPLUS_IMAGE_PREFIX-}langchainplus-backend:latest
-    environment:
-      - PORT=1984
-      - LANGCHAIN_ENV=local_docker
-      - LOG_LEVEL=warning
-      - OPENAI_API_KEY=${OPENAI_API_KEY}
-    ports:
-      - 1984:1984
-    depends_on:
-      - langchain-db
-    build:
-      context: backend/.
-      dockerfile: Dockerfile
-  langchain-db:
-    image: postgres:14.1
-    command:
-      [
-        "postgres",
-        "-c",
-        "log_min_messages=WARNING",
-        "-c",
-        "client_min_messages=WARNING"
-      ]
-    environment:
-      - POSTGRES_PASSWORD=postgres
-      - POSTGRES_USER=postgres
-      - POSTGRES_DB=postgres
-    volumes:
-      - langchain-db-data:/var/lib/postgresql/data
-    ports:
-      - 5433:5432
-volumes:
-  langchain-db-data:
