+++ langchain/client/__init__.py
@@ -1,6 +1,9 @@
 """LangChain+ Client."""
+from langchain.client.runner_utils import (
+    arun_on_dataset,
+    arun_on_examples,
+    run_on_dataset,
+    run_on_examples,
+)
 
-
-from langchain.client.langchain import LangChainPlusClient
-
-__all__ = ["LangChainPlusClient"]
+__all__ = ["arun_on_dataset", "run_on_dataset", "arun_on_examples", "run_on_examples"]
