+++ pyproject.toml
@@ -9,7 +9,6 @@ repository = "https://www.github.com/hwchase17/langchain"
 
 [tool.poetry.scripts]
 langchain-server = "langchain.server:main"
-langchain = "langchain.cli.main:main"
 
 [tool.poetry.dependencies]
 python = ">=3.8.1,<4.0"
@@ -104,6 +103,7 @@ momento = {version = "^1.5.0", optional = true}
 bibtexparser = {version = "^1.4.0", optional = true}
 pyspark = {version = "^3.4.0", optional = true}
 tigrisdb = {version = "^1.0.0b6", optional = true}
+langchainplus-sdk = "^0.0.4"
 
 [tool.poetry.group.docs.dependencies]
 autodoc_pydantic = "^1.8.0"
